feat: refactored separate replace and create mUSD conversion tx helpers into single file
